fix(Storybook): moves control rows into components folder
